Divide Logic into a base Controller
